Fix fh-run-array to take string or array again
JIRA: https://issues.jboss.org/browse/RHMAP-5713

Motivation:
It turns out that when we moved away from `grunt.template.process` to
`grunt.config.get` in 1.0.0, it returns a non-string, if it finds a
non-string. Useful, but not what we were expecting! :)

Modification:
Only split if it's a string, otherwise assume it's an array with
multiple commands, and join with ' && '.

Result:
Single-command strings, and multi-command arrays work again.